fix testing
