--- FIX/fix/resources/prime.properties
@@ -1,8 +1,0 @@
-[session_1]
-passphrase = YOUR_API_PASSPHRASE
-api_key = YOUR_API_KEY
-api_secret = YOUR_API_SECRET
-portfolio = YOUR_PORTFOLIO_ID
-fix_version = FIX.4.2
-sender_comp_id = YOUR_SVC_ACCOUNT_ID
-target_comp_id = COIN
